lr4-backprop/requirements.txt
colorama==0.4.6
numpy==2.1.2
pillow==11.0.0
